+++ src/errors/instances.ts
@@ -8,7 +8,7 @@ import {
     BitError,
     MetadataError,
     SettingError,
-    EVMCallError
+    EVMCallError,
 } from './errors';
 
 export const emptyContractError = new ContractError('unexpected empty contract code of given jetton.', 100);
@@ -35,4 +35,9 @@ export const emptySettingError = (setting: string) =>
     new SettingError(`unexpected empty ${setting}. Make sure the settings contract is valid.`, 110);
 
 export const invalidMethodNameError = (methodName: string) =>
-    new EVMCallError(`Invalid Solidity method name: "${methodName}". Method must be either a valid identifier or have parameters (bytes,bytes).`, 110);
+    new EVMCallError(
+        `Invalid Solidity method name: "${methodName}". Method must be either a valid identifier or have parameters (bytes,bytes).`,
+        111,
+    );
+
+export const simulationError = new FetchError('Failed to simulate EVM call', 112);