+++ src/sdk/TacSdk.ts
@@ -1,3 +1,4 @@
+import axios from 'axios';
 import { Address, address, beginCell, Cell, toNano } from '@ton/ton';
 import { ethers, keccak256, toUtf8Bytes, isAddress as isEthereumAddress } from 'ethers';
 import type { SenderAbstraction } from '../sender';
@@ -12,6 +13,8 @@ import {
     TACParams,
     RawAssetBridgingData,
     UserWalletBalanceExtended,
+    EVMSimulationResults,
+    EVMSimulationRequest,
 } from '../structs/Struct';
 // import internal structs
 import {
@@ -23,17 +26,14 @@ import {
     AssetOpType,
     ShardMessage,
     ShardTransaction,
+    EVMSimulationResponse,
 } from '../structs/InternalStruct';
 // jetton imports
 import { JettonMaster } from '../wrappers/JettonMaster';
 import { JettonWallet } from '../wrappers/JettonWallet';
 // ton settings
 import { Settings } from '../wrappers/Settings';
-import {
-    JETTON_TRANSFER_FORWARD_TON_AMOUNT,
-    TRANSACTION_TON_AMOUNT,
-    DEFAULT_DELAY,
-} from './Consts';
+import { JETTON_TRANSFER_FORWARD_TON_AMOUNT, TRANSACTION_TON_AMOUNT, DEFAULT_DELAY } from './Consts';
 import {
     buildEvmDataCell,
     calculateAmount,
@@ -42,12 +42,13 @@ import {
     calculateRawAmount,
     generateRandomNumberByTimestamp,
     generateTransactionLinker,
+    toCamelCaseTransformer,
     sleep,
     validateEVMAddress,
     validateTVMAddress,
 } from './Utils';
 import { mainnet, testnet } from '@tonappchain/artifacts';
-import { emptyContractError } from '../errors';
+import { emptyContractError, simulationError } from '../errors';
 import { orbsOpener4 } from '../adapters/contractOpener';
 
 export class TacSdk {
@@ -56,6 +57,7 @@ export class TacSdk {
     readonly artifacts: typeof testnet | typeof mainnet;
     readonly TONParams: InternalTONParams;
     readonly TACParams: InternalTACParams;
+    readonly liteSequencerEndpoints: string[];
 
     private constructor(
         network: Network,
@@ -63,12 +65,14 @@ export class TacSdk {
         artifacts: typeof testnet | typeof mainnet,
         TONParams: InternalTONParams,
         TACParams: InternalTACParams,
+        liteSequencerEndpoints: string[],
     ) {
         this.network = network;
         this.delay = delay;
         this.artifacts = artifacts;
         this.TONParams = TONParams;
         this.TACParams = TACParams;
+        this.liteSequencerEndpoints = liteSequencerEndpoints;
     }
 
     static async create(sdkParams: SDKParams): Promise<TacSdk> {
@@ -77,7 +81,12 @@ export class TacSdk {
         const artifacts = network === Network.Testnet ? testnet : mainnet;
         const TONParams = await this.prepareTONParams(network, delay, artifacts, sdkParams.TONParams);
         const TACParams = await this.prepareTACParams(artifacts, sdkParams.TACParams);
-        return new TacSdk(network, delay, artifacts, TONParams, TACParams);
+        const liteSequencerEndpoints =
+            sdkParams.customLiteSequencerEndpoints ??
+            (network === Network.Testnet
+                ? testnet.PUBLIC_LITE_SEQUENCER_ENDPOINTS
+                : mainnet.PUBLIC_LITE_SEQUENCER_ENDPOINTS);
+        return new TacSdk(network, delay, artifacts, TONParams, TACParams, liteSequencerEndpoints);
     }
 
     private static async prepareTONParams(
@@ -112,9 +121,7 @@ export class TacSdk {
         artifacts: typeof testnet | typeof mainnet,
         TACParams?: TACParams,
     ): Promise<InternalTACParams> {
-        const provider =
-            TACParams?.provider ??
-            ethers.getDefaultProvider(artifacts.TAC_RPC_ENDPOINT);
+        const provider = TACParams?.provider ?? ethers.getDefaultProvider(artifacts.TAC_RPC_ENDPOINT);
 
         const settingsAddress = TACParams?.settingsAddress?.toString() ?? artifacts.tac.addresses.TAC_SETTINGS_ADDRESS;
         const settings = new ethers.Contract(
@@ -505,5 +512,24 @@ export class TacSdk {
         });
 
         return jettonMaster.address.toString();
+    }
+
+    async simulateEVMMessage(req: EVMSimulationRequest): Promise<EVMSimulationResults> {
+        for (const endpoint of this.liteSequencerEndpoints) {
+            try {
+                const response = await axios.post<EVMSimulationResponse>(
+                    `${endpoint}/evm/simulator/simulate-message`,
+                    req,
+                    {
+                        transformResponse: [toCamelCaseTransformer],
+                    },
+                );
+
+                return response.data.response;
+            } catch (error) {
+                console.error(`Error while simulating with ${endpoint}:`, error);
+            }
+        }
+        throw simulationError;
     }
 }